backend/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
pdfplumber
Flask-Login
gunicorn
